install: fix where to search for config dirs
When running from outsinde the `./install` directory, the `link` script
was taking the root directory as "the source of config directories",
hence copying everyting from its parent insad of the parent of
`./install` (the root directory).

Now, the `link` script will check if it's being invoked from within the
`./install` directory or the root of the project and decide where to
grab all the configuration directories from.